requirements/common.txt
celery
django
django-tastypie
django-redis
django-environ
gunicorn
psycopg2
redis
requests
requests_oauthlib
